Squashed commit of the following:

commit 888682bccf55b3b6f93c6d2a023fc295e34b99d6
Author: jrittle <jrittle@cisco.com>
Date:   Wed Feb 24 09:40:49 2021 -0500

    iec104: additional input sanitization, syntax, and style changes

+++ src/service_inspectors/iec104/iec104_parse_information_object_elements.h
@@ -46,7 +46,7 @@ void parseIec104Srq(const Iec104SrqType* srq);
 void parseIec104Scq(const Iec104ScqType* scq);
 void parseIec104Lsq(const Iec104LsqType* lsq);
 void parseIec104Afq(const Iec104AfqType* afq);
-uint8_t parseIec104Vsq(const Iec104ApciI* apci);
+uint32_t parseIec104Vsq(const Iec104ApciI* apci);
 void parseIec104CauseOfTx(const Iec104ApciI* apci);
 void parseIec104TwoOctetCommonAddress(const Iec104ApciI* apci);
 void parseIec104InformationObjectAddressWithThreeOctets(
